requirements.txt
flask
piexif
pillow
pyyaml